Support Connect message
When reconnecting to a different shell to the one that created it we
need to support the Connect protocol message as well as call the Connect
method in powershell. Unlike Reconnect, Connect needs to reestablish
client-side state which is why it is different from Reconnect that
already has that state as a result of the initial CreateShell.