Updated shell shutdown logic
Update the shell shutdown notification logic into powershell and handle
the operationcomplete callbacks properly. Also fixed the RefuseUnload
and RequestUnload calls so they are done on the Shell context instead of
the Load context -- the Load context is temporary and the lifetime of it
went away, The operation timeout can be made more long-lived.