error handling, create shell sends a request off box
Errors messages that are generated now get propagated to client so
it is easier to troubleshoot.
The CreateShell instance is sent off the box and the response (currently
an error) is propagated back to the client. WSMan layer now needs to
package the instance properly, and pass the correct resource uri in the
headers.