Latest changes for PSRP Client
This is not release quality yet. It supports powershell on Linux to
connect to a Windows box and create a shell and run stuff.
From this shell you should be able to do most things but this is an
early release so don't file bugs against me yet!
Currently you need a local admin account on the Windows box and you need
to connect using basic authentication with this local machine account.
You cannot pipe anything into the shell yet. You cannot
disconnect/reconnect yet. The data stream is not compressed yet. Linux
to Linux is broken for some odd reason that I am currently
investigating. There are memory leaks. PowerShell cannot be closed after
a remote shell is created even though you can exit from the shell and do
things locally, PowerShell is blocked from shutting down because some
handle has not been cancelled yet. Again, working on it!